requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pytest
pytest-asyncio
google-generativeai
httpx
reportlab
requests
jwt
python-dotenv
aiohttp